Move setup/References.kt into connection/Messages.kt
