+++ common/src/main/java/com/boswelja/devicemanager/common/connection/Messages.kt
@@ -12,4 +12,8 @@ object Messages {
 
     const val REQUEST_APP_VERSION = "/request_app_version"
     const val REQUEST_UPDATE_CAPABILITIES = "/update_capabilities"
+
+    const val CHECK_WATCH_REGISTERED_PATH = "/check_watch_registered"
+    const val WATCH_NOT_REGISTERED_PATH = "/watch_not_registered"
+    const val WATCH_REGISTERED_PATH = "/watch_registered"
 }
