--- common/src/main/java/com/boswelja/devicemanager/common/setup/References.kt
@@ -1,15 +1,0 @@
-/* Copyright (C) 2020 Jack Boswell <boswelja@outlook.com>
- *
- * This file is part of Wearable Extensions
- *
- * This file, and any part of the Wearable Extensions app/s cannot be copied and/or distributed
- * without permission from Jack Boswell (boswelja) <boswela@outlook.com>
- */
-package com.boswelja.devicemanager.common.setup
-
-object References {
-
-    const val CHECK_WATCH_REGISTERED_PATH = "/check_watch_registered"
-    const val WATCH_NOT_REGISTERED_PATH = "/watch_not_registered"
-    const val WATCH_REGISTERED_PATH = "/watch_registered"
-}
