+++ mobile/src/main/java/com/boswelja/devicemanager/watchmanager/connection/wearos/WatchMessageReceiver.kt
@@ -16,14 +16,14 @@ import com.boswelja.devicemanager.common.Compat
 import com.boswelja.devicemanager.common.Extensions.toByteArray
 import com.boswelja.devicemanager.common.batterysync.References.REQUEST_BATTERY_UPDATE_PATH
 import com.boswelja.devicemanager.common.batterysync.Utils
+import com.boswelja.devicemanager.common.connection.Messages.CHECK_WATCH_REGISTERED_PATH
 import com.boswelja.devicemanager.common.connection.Messages.LAUNCH_APP
 import com.boswelja.devicemanager.common.connection.Messages.LOCK_PHONE
+import com.boswelja.devicemanager.common.connection.Messages.WATCH_NOT_REGISTERED_PATH
+import com.boswelja.devicemanager.common.connection.Messages.WATCH_REGISTERED_PATH
 import com.boswelja.devicemanager.common.connection.References.CAPABILITY_WATCH_APP
 import com.boswelja.devicemanager.common.dndsync.References.REQUEST_INTERRUPT_FILTER_ACCESS_STATUS_PATH
 import com.boswelja.devicemanager.common.preference.PreferenceKey
-import com.boswelja.devicemanager.common.setup.References.CHECK_WATCH_REGISTERED_PATH
-import com.boswelja.devicemanager.common.setup.References.WATCH_NOT_REGISTERED_PATH
-import com.boswelja.devicemanager.common.setup.References.WATCH_REGISTERED_PATH
 import com.boswelja.devicemanager.common.ui.activity.BaseWatchPickerPreferenceActivity.Companion.EXTRA_PREFERENCE_KEY
 import com.boswelja.devicemanager.dndsync.ui.DnDSyncPreferenceActivity
 import com.boswelja.devicemanager.main.MainActivity
