+++ wearos/src/main/java/com/boswelja/devicemanager/onboarding/ui/OnboardingViewModel.kt
@@ -17,8 +17,8 @@ import androidx.lifecycle.Transformations
 import androidx.preference.PreferenceManager
 import com.boswelja.devicemanager.R
 import com.boswelja.devicemanager.common.Event
-import com.boswelja.devicemanager.common.setup.References
-import com.boswelja.devicemanager.common.setup.References.CHECK_WATCH_REGISTERED_PATH
+import com.boswelja.devicemanager.common.connection.Messages.CHECK_WATCH_REGISTERED_PATH
+import com.boswelja.devicemanager.common.connection.Messages.WATCH_REGISTERED_PATH
 import com.boswelja.devicemanager.phoneconnectionmanager.References.PHONE_ID_KEY
 import com.google.android.gms.wearable.MessageClient
 import com.google.android.gms.wearable.NodeClient
@@ -38,7 +38,7 @@ constructor(
     val messageListener =
         MessageClient.OnMessageReceivedListener {
             when (it.path) {
-                References.WATCH_REGISTERED_PATH -> {
+                WATCH_REGISTERED_PATH -> {
                     sharedPreferences.edit { putString(PHONE_ID_KEY, it.sourceNodeId) }
                     onWatchRegistered.fire()
                 }
